Moved XML parser code to Request class  - resulting data from request has yet to be communicated back to the user  - lacking check of online status
